+++ demos/demo40/gen.sh
@@ -1,0 +1,4 @@
+cp ../demo1/temperature.h .
+node ../../../awtk/tools/idl_gen/index.js idl.json temperature.h
+node ../../tools/view_model_gen/gen_vm.js idl.json
+