+++ demos/demo40/temperature_converter.h
@@ -1,0 +1,33 @@
+﻿/**
+ * File:   temperature_converter.h
+ * Author: AWTK Develop Team
+ * Brief:  temperature converter
+ *
+ * Copyright (c) 2019 - 2020  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2019-02-02 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#ifndef TK_TEMPERATURE_CONVERTER_H
+#define TK_TEMPERATURE_CONVERTER_H
+
+#include "mvvm/base/view_model.h"
+
+BEGIN_C_DECLS
+
+ret_t temperature_converter_init(void);
+
+END_C_DECLS
+
+#endif /*TK_TEMPERATURE_CONVERTER_H*/