+++ design/default/scripts/temperature_ex.js
@@ -10,6 +10,28 @@ ViewModel('temperature_ex', {
     apply: function(args) {
       this.applydValue = this.value;
       this.notifyPropsChanged();
+    },
+    canChangeLocale: function() {
+      return true;
+    },
+    changeLocale: function() {
+      locale = getLocale();
+      if (locale.language == 'zh') {
+        setLocale('en', 'US');
+      } else {
+        setLocale('zh', 'CN');
+      }
+    },
+    canChangeTheme: function() {
+      return true;
+    },
+    changeTheme: function() {
+      theme = getTheme();
+      if (theme == 'default') {
+        setTheme('dark');
+      } else {
+        setTheme('default');
+      }
     }
   },
   onWillMount: function(req) {